+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Current Trunk
 - HeapType::ext has been restored but is no longer a subtype of HeapType::any to
   match the latest updates in the GC spec. (#4898)
 - `i31ref` and `dataref` are now nullable to match the latest GC spec. (#4843)
+- Add support for `extern.externalize` and `extern.internalize`. (#4975)
 
 v109
 ----